Update and rename lab2.py to main_lab2.py
